requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.3
click==8.1.6
cmake==3.27.0
coloredlogs==15.0.1
cryptography==41.0.2
dataclasses-json==0.5.13
decorator==4.4.2
dnspython==2.4.1
elevenlabs==0.2.21
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.99.1
ffmpeg-python==0.2.0
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
future==0.18.3
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-resources==6.0.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
langchain==0.0.247
langsmith==0.0.15
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==10.0.0
moviepy==1.0.3
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nanoid==2.0.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numexpr==2.8.4
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openai-whisper==20230314
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
posthog==3.0.1
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.23.4
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pydub==0.25.1
Pygments==2.15.1
pymongo==4.4.1
pypandoc==1.11
PyPika==0.48.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.3.1
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.8.7
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
